requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astunparse==1.6.3
attrs==19.1.0
blis==0.4.1
boto3==1.12.32
botocore==1.15.49
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.3
distlib==0.3.1
docutils==0.15.2
filelock==3.0.12
Flask==1.1.1
Flask-API==2.0
Flask-WTF==0.14.3
gast==0.3.3
google-api-core==1.22.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-storage==1.28.1
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
htmlmin==0.1.12
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.10.0
jsonschema==3.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mixpanel==4.4.0
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.3.0.38
opt-einsum==3.3.0
orjson==3.3.0
pandas==1.1.0
Pillow==6.2.1
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
rarfile==4.0
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
rq==1.5.0
rsa==4.6
s3transfer==0.3.3
scipy==1.4.1
six==1.12.0
srsly==0.2.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tqdm==4.40.2
urllib3==1.25.7
virtualenv==20.0.30
virtualenvwrapper-win==1.2.6
wasabi==0.4.2
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.3.3
xmljson==0.2.0

